Add files via upload
There are two parts: The LCDbackpackMotorScaleV4 resides on the Arduino Uno and contains the switch, LCD, and stepper motor code.  Because I couldn't overcome blocking issues between the scale and the switches, I wrote separate code, OLEDdisplayPart to sit on a nano.  This code monitors the scale and displays units on the OLED screen.